src/requirements.txt
Django==1.11.22
django-crispy-forms==1.7.2
pytz==2019.1
